Refactor: property 교체를 위한 패키지 변경
