+++ src/main/kotlin/code_immotion/server/application/controller/PropertyController.kt
@@ -1,7 +1,7 @@
 package code_immotion.server.application.controller
 
-import code_immotion.server.domain.property.PropertyFacade
-import code_immotion.server.domain.property.dto.PropertyCondition
+import code_immotion.server.domain.old_property.PropertyFacade
+import code_immotion.server.domain.old_property.dto.PropertyCondition
 import io.github.oshai.kotlinlogging.KotlinLogging
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.Parameter
