+++ src/main/kotlin/code_immotion/server/domain/old_property/PropertyFacade.kt
@@ -1,11 +1,9 @@
-package code_immotion.server.domain.property
+package code_immotion.server.domain.old_property
 
-import code_immotion.server.application.handler.exception.CustomException
-import code_immotion.server.application.handler.exception.ErrorCode
 import code_immotion.server.domain.open_api.client.OpenApiClient
-import code_immotion.server.domain.property.dto.PropertyAggregation
-import code_immotion.server.domain.property.dto.PropertyCondition
-import code_immotion.server.domain.property.dto.PropertyResponse
+import code_immotion.server.domain.old_property.dto.PropertyAggregation
+import code_immotion.server.domain.old_property.dto.PropertyCondition
+import code_immotion.server.domain.old_property.dto.PropertyResponse
 import org.springframework.data.geo.Point
 import org.springframework.stereotype.Component
