+++ src/main/kotlin/code_immotion/server/domain/old_property/PropertyService.kt
@@ -1,11 +1,11 @@
-package code_immotion.server.domain.property
+package code_immotion.server.domain.old_property
 
 import code_immotion.server.application.handler.exception.CustomException
 import code_immotion.server.application.handler.exception.ErrorCode
-import code_immotion.server.domain.property.dto.PropertyAggregation
-import code_immotion.server.domain.property.dto.PropertyCondition
-import code_immotion.server.domain.property.dto.PropertyResponse
-import code_immotion.server.domain.property.entity.Property
+import code_immotion.server.domain.old_property.dto.PropertyAggregation
+import code_immotion.server.domain.old_property.dto.PropertyCondition
+import code_immotion.server.domain.old_property.dto.PropertyResponse
+import code_immotion.server.domain.old_property.entity.Property
 import io.github.oshai.kotlinlogging.KotlinLogging
 import org.springframework.data.geo.Point
 import org.springframework.stereotype.Service
