+++ src/main/kotlin/code_immotion/server/domain/old_property/dto/PropertyAggregation.kt
@@ -1,4 +1,4 @@
-package code_immotion.server.domain.property.dto
+package code_immotion.server.domain.old_property.dto
 
 import io.swagger.v3.oas.annotations.media.Schema
