+++ src/main/kotlin/code_immotion/server/domain/old_property/dto/PropertyResponse.kt
@@ -1,8 +1,8 @@
-package code_immotion.server.domain.property.dto
+package code_immotion.server.domain.old_property.dto
 
-import code_immotion.server.domain.property.entity.HouseType
-import code_immotion.server.domain.property.entity.Property
-import code_immotion.server.domain.property.entity.TradeType
+import code_immotion.server.domain.old_property.entity.HouseType
+import code_immotion.server.domain.old_property.entity.Property
+import code_immotion.server.domain.old_property.entity.TradeType
 import io.swagger.v3.oas.annotations.media.Schema
 import org.springframework.data.geo.Distance
 import org.springframework.data.geo.Point
