+++ src/main/kotlin/code_immotion/server/domain/old_property/dto/SortType.kt
@@ -1,4 +1,4 @@
-package code_immotion.server.domain.property.dto
+package code_immotion.server.domain.old_property.dto
 
 enum class SortType(val value: String) {
     PRICE(PropertyAggregation.Data::averagePrice.name), DISTANCE(PropertyAggregation.Data::averageDistance.name)
