+++ src/main/kotlin/code_immotion/server/domain/old_property/entity/DepositRent.kt
@@ -1,9 +1,9 @@
-package code_immotion.server.domain.property.entity
+package code_immotion.server.domain.old_property.entity
 
 import com.fasterxml.jackson.databind.JsonNode
 import java.time.LocalDate
 
-class Sale(
+class DepositRent(
     id: String? = null,
     buildingName: String,
     address: String,
@@ -17,14 +17,14 @@ class Sale(
     exclusiveArea: Int,
 ) : Property(id, buildingName, address, addressNumber, houseType, tradeType, floor, price, null, dealDate, buildYear, exclusiveArea) {
     companion object {
-        fun from(jsonNode: JsonNode, state: String, city: String, houseType: HouseType, buildingName: String) = Sale(
+        fun from(jsonNode: JsonNode, state: String, city: String, houseType: HouseType, buildingName: String) = DepositRent(
             buildingName = buildingName,
             address = "$state $city ${jsonNode.path("umdNm").asText()}",
             addressNumber = jsonNode.path("jibun").asText(),
             houseType = houseType,
-            tradeType = TradeType.SALE,
+            tradeType = TradeType.LONG_TERM_RENT,
             floor = jsonNode.path("floor").asInt(),
-            price = jsonNode.path("dealAmount").asText().replace(",", "").toLong(),
+            price = jsonNode.path("deposit").asText().replace(",", "").toLong(),
             buildYear = jsonNode.path("buildYear").asInt(),
             exclusiveArea = jsonNode.path("excluUseAr").asDouble().toInt(),
             dealDate = LocalDate.of(
