+++ src/main/kotlin/code_immotion/server/domain/old_property/entity/Sale.kt
@@ -1,9 +1,9 @@
-package code_immotion.server.domain.property.entity
+package code_immotion.server.domain.old_property.entity
 
 import com.fasterxml.jackson.databind.JsonNode
 import java.time.LocalDate
 
-class DepositRent(
+class Sale(
     id: String? = null,
     buildingName: String,
     address: String,
@@ -17,14 +17,14 @@ class DepositRent(
     exclusiveArea: Int,
 ) : Property(id, buildingName, address, addressNumber, houseType, tradeType, floor, price, null, dealDate, buildYear, exclusiveArea) {
     companion object {
-        fun from(jsonNode: JsonNode, state: String, city: String, houseType: HouseType, buildingName: String) = DepositRent(
+        fun from(jsonNode: JsonNode, state: String, city: String, houseType: HouseType, buildingName: String) = Sale(
             buildingName = buildingName,
             address = "$state $city ${jsonNode.path("umdNm").asText()}",
             addressNumber = jsonNode.path("jibun").asText(),
             houseType = houseType,
-            tradeType = TradeType.LONG_TERM_RENT,
+            tradeType = TradeType.SALE,
             floor = jsonNode.path("floor").asInt(),
-            price = jsonNode.path("deposit").asText().replace(",", "").toLong(),
+            price = jsonNode.path("dealAmount").asText().replace(",", "").toLong(),
             buildYear = jsonNode.path("buildYear").asInt(),
             exclusiveArea = jsonNode.path("excluUseAr").asDouble().toInt(),
             dealDate = LocalDate.of(
