+++ src/main/kotlin/code_immotion/server/domain/old_property/entity/TradeType.kt
@@ -1,4 +1,4 @@
-package code_immotion.server.domain.property.entity
+package code_immotion.server.domain.old_property.entity
 
 enum class TradeType {
     SALE, MONTHLY_RENT, LONG_TERM_RENT
